+++ src/test/java/com/jupiter/tools/spring/test/mongo/idea/skip/test/IdeaSkipTestRule.java
@@ -1,4 +1,4 @@
-package com.jupiter.tools.spring.test.mongo.junit4;
+package com.jupiter.tools.spring.test.mongo.idea.skip.test;
 
 import org.junit.AssumptionViolatedException;
 import org.junit.rules.TestRule;
@@ -6,7 +6,11 @@ import org.junit.runner.Description;
 import org.junit.runners.model.Statement;
 
 
-public class IdeaTestSkipRule implements TestRule {
+/**
+ * This rule provides an ability to skip
+ * the junit4 test when this test running from IDEA
+ */
+public class IdeaSkipTestRule implements TestRule {
 
     private boolean isRunInIdea() {
         String property = System.getProperty("sun.java.command");
