+++ src/test/java/com/jupiter/tools/spring/test/mongo/junit4/MongoDbRuleExpectedIT.java
@@ -4,9 +4,10 @@ import com.jupiter.tools.spring.test.mongo.Bar;
 import com.jupiter.tools.spring.test.mongo.Foo;
 import com.jupiter.tools.spring.test.mongo.annotation.ExpectedMongoDataSet;
 import com.jupiter.tools.spring.test.mongo.annotation.MongoDataSet;
+import com.jupiter.tools.spring.test.mongo.idea.skip.test.IdeaSkipTest;
+import com.jupiter.tools.spring.test.mongo.idea.skip.test.IdeaSkipTestRule;
 import org.junit.ClassRule;
 import org.junit.Test;
-import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
 import org.junit.platform.launcher.LauncherDiscoveryRequest;
 import org.junit.platform.launcher.core.LauncherFactory;
@@ -30,7 +31,7 @@ import static org.junit.platform.testkit.engine.TestExecutionResultConditions.me
  *
  * @author Korovin Anatoliy
  */
-@DisabledIfSystemProperty(named = "sun.java.command", matches = "com.intellij.rt.execution.*")
+@IdeaSkipTest
 public class MongoDbRuleExpectedIT {
 
     @org.junit.jupiter.api.Test
@@ -147,7 +148,7 @@ public class MongoDbRuleExpectedIT {
         private MongoTemplate mongoTemplate;
 
         @ClassRule
-        public static IdeaTestSkipRule ideaTestSkipRule = new IdeaTestSkipRule();
+        public static IdeaSkipTestRule ideaSkipTestRule = new IdeaSkipTestRule();
 
         @Test
         @MongoDataSet(cleanAfter = true, cleanBefore = true)
